[EasyWebhook] Test support for illuminate/database
