+++ packages/EasyWebhook/composer.json
@@ -16,6 +16,7 @@
     "require-dev": {
         "doctrine/dbal": "^2.6",
         "eonx-com/easy-pagination": "^3.1",
+        "illuminate/database": "^5.5 || ^6.0",
         "ramsey/uuid": "^3.9",
         "phpunit/phpunit": "^8.4 || ^9.5",
         "laravel/lumen-framework": "^5.5",
