+++ packages/EasyWebhook/tests/AbstractStoreTestCase.php
@@ -6,6 +6,7 @@ namespace EonX\EasyWebhook\Tests;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DriverManager;
+use EonX\EasyWebhook\Bridge\Doctrine\StatementProviders\SqliteStatementProvider;
 use EonX\EasyWebhook\Interfaces\WebhookInterface;
 use EonX\EasyWebhook\Webhook;
 
@@ -14,7 +15,7 @@ abstract class AbstractStoreTestCase extends AbstractTestCase
     /**
      * @var \Doctrine\DBAL\Connection
      */
-    private $doctrineDbal;
+    protected $doctrineDbal;
 
     protected function createWebhookForSendAfter(
         ?\DateTimeInterface $sendAfter = null,
@@ -45,5 +46,30 @@ abstract class AbstractStoreTestCase extends AbstractTestCase
         return $this->doctrineDbal = DriverManager::getConnection([
             'url' => 'sqlite:///:memory:',
         ]);
+    }
+
+    protected function setUp(): void
+    {
+        $conn = $this->getDoctrineDbalConnection();
+        $conn->connect();
+
+        foreach (SqliteStatementProvider::migrateStatements() as $statement) {
+            $conn->executeStatement($statement);
+        }
+
+        parent::setUp();
+    }
+
+    protected function tearDown(): void
+    {
+        $conn = $this->getDoctrineDbalConnection();
+
+        foreach (SqliteStatementProvider::rollbackStatements() as $statement) {
+            $conn->executeStatement($statement);
+        }
+
+        $conn->close();
+
+        parent::tearDown();
     }
 }
