+++ packages/EasyWebhook/tests/Stores/DoctrineDbalResultStoreTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace EonX\EasyWebhook\Tests\Stores;
 
-use EonX\EasyWebhook\Bridge\Doctrine\StatementProviders\SqliteStatementProvider;
 use EonX\EasyWebhook\Interfaces\WebhookInterface;
 use EonX\EasyWebhook\Stores\DoctrineDbalResultStore;
 use EonX\EasyWebhook\Tests\AbstractStoreTestCase;
@@ -23,30 +22,5 @@ final class DoctrineDbalResultStoreTest extends AbstractStoreTestCase
         $store->store($result);
 
         self::assertNotEmpty($result->getId());
-    }
-
-    protected function setUp(): void
-    {
-        $conn = $this->getDoctrineDbalConnection();
-        $conn->connect();
-
-        foreach (SqliteStatementProvider::migrateStatements() as $statement) {
-            $conn->executeStatement($statement);
-        }
-
-        parent::setUp();
-    }
-
-    protected function tearDown(): void
-    {
-        $conn = $this->getDoctrineDbalConnection();
-
-        foreach (SqliteStatementProvider::rollbackStatements() as $statement) {
-            $conn->executeStatement($statement);
-        }
-
-        $conn->close();
-
-        parent::tearDown();
     }
 }
